postprocess: Support multiple named transforms
Adds base class for transforms.
